+++ docs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Package `simplego`:
   * Added `GetBackend` that returns a singleton backend, created with the default configuration at the first request.
+* Package `ui/commandline`:
+  * Added optional extra arbitrary metrics to print in the command-line with `AttachProgressBar`.
+  * Added `FormatDuration` to pretty-print duration.
 
 # v0.20.1: 2025/06/12 Trainer.AccumulateGradients (when the batch doesn't fit memory); VNN fixes; Numpy improvements.